Refactor JDV source paths Remove <Router>
