+++ src/web/jdv/client/src/model/Plot.ts
